requirements.txt
openai
qdrant-client
supabase
python-dotenv
openai-agents
fastapi
uvicorn
gunicorn
pydantic
requests
